+++ src/checkDbExistence.js
@@ -4,18 +4,16 @@ async function checkDbExistence(config) {
     let result = false;
     
     if (!config.cliMode) {
-        config.debug(`Checking master database ${chalk.magenta(config.database.database)} ...`)
+        config.debug(`Checking database ${chalk.magenta(config.database.database)} existence ...`)
 
         try {
             await config.db.dbExists(config.database.database);
 
-            config.debug(`database exists`)
+            config.debug(`database exists`);
 
             result = true
         } catch (ex) {
-            console.error(chalk.red(`Master database does not exist or cannot check its existence.
-Operation aborted.`));
-            config.debug(ex);
+            config.error = ex;
         }
     } else {
         result = true;