+++ src/services/DbHelper/exceptions/ConnectionException.js
@@ -1,0 +1,6 @@
+import { Exception } from "@locustjs/exception";
+
+class ConnectionException extends Exception {
+}
+
+export default ConnectionException;